Moving components to the Sidebar folder and creating and importing the sidebarData array
