+++ src/Components/Backoffice/Sidebar/sidebarData.js
@@ -1,0 +1,34 @@
+export const sidebarData = [
+    {
+        section: 'Actividades',
+        path: '/backoffice/activities',
+    },
+    {
+        section: 'Categorías',
+        path: '/backoffice/categories',
+    },
+    {
+        section: 'Home',
+        path: '/backoffice/home',
+    },
+    {
+        section: 'Miembros',
+        path: '/backoffice/members',
+    },
+    {
+        section: 'Novedades',
+        path: '/backoffice/news/',
+    },
+    {
+        section: 'Organización',
+        path: '/backoffice/organization',
+    },
+    {
+        section: 'Slides',
+        path: '/backoffice/slides',
+    },
+    {
+        section: 'Usuarios',
+        path: '/backoffice/users',
+    }
+];
